+++ include/core/game/modifier/SpeedBoostModifier.h
@@ -1,12 +1,14 @@
 #pragma once
 
+#include <cstdint>
+
 #include "core/game/modifier/TimedModifier.h"
 #include "core/game/physics/PObject.h"
 #include "core/math/vector.h"
 
 struct SpeedBoostModifierData : TimedModifierData {
   float factorIncrease;
-  SpeedBoostModifierData(unsigned long long expire, float factorIncrease)
+  SpeedBoostModifierData(uint64_t expire, float factorIncrease)
       : TimedModifierData(expire), factorIncrease(factorIncrease) {}
 };
 class SpeedBoostModifier : public TimedModifier {
